# backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Amana Backend API",
    "version": "1.0.0",
    "description": "OpenAPI coverage for the live Amana backend routes under `backend/src/routes`.\n\nAuthentication flow:\n1. Call `POST /auth/challenge` with a Stellar public key.\n2. Sign the returned challenge with the wallet.\n3. Exchange the signature at `POST /auth/verify` for a bearer token.\n4. Send `Authorization: Bearer <token>` to protected endpoints.\n\nJWT semantics enforced by middleware:\n- `iss` must match `JWT_ISSUER` (default `amana`)\n- `aud` must match `JWT_AUDIENCE` (default `amana-api`)\n- `jti` is required and checked against the revocation denylist\n- `nbf` must be present and not be in the future\n\nError code references in operation descriptions map to backend `ErrorCode` values in\n`backend/src/errors/errorCodes.ts`. Some route handlers still emit legacy `{ error }`\npayloads while centralized validation/runtime failures return `{ code, message, details }`."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "TradeIdPath": {
        "in": "path",
        "name": "id",
        "required": true,
        "description": "Trade identifier used by the backend route layer.",
        "schema": {
          "type": "string"
        },
        "example": 4294967297
      },
...
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "example": {
          "error": "Unauthorized"
        }
      },
      "AppErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "AUTH_ERROR",
              "DOMAIN_ERROR",
              "INFRA_ERROR",
              "NOT_FOUND",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "oneOf": [
              {
                "type": "object"
              },
              {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            ]
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "example": {
          "code": "VALIDATION_ERROR",
          "message": "Validation failed",
          "details": [
            {
              "path": "body.amountUsdc",
              "message": "Invalid amount format"
            }
          ]
        }
      },
      "AuthChallengeRequest": {
        "type": "object",
        "properties": {
          "walletAddress": {
            "type": "string",
            "description": "Stellar Ed25519 public key."
          }
        },
        "required": [
          "walletAddress"
        ],
        "example": {
          "walletAddress": "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF"
        }
      },
      "AuthChallengeResponse": {
        "type": "object",
        "properties": {
          "challenge": {
            "type": "string"
          }
        },
        "required": [
          "challenge"
        ],
        "example": {
          "challenge": "amana:login:1742794421:7ced1c65a9a44a7d"
        }
      },
      "AuthVerifyRequest": {
        "type": "object",
        "properties": {
          "walletAddress": {
            "type": "string"
          },
          "signedChallenge": {
            "type": "string",
            "description": "Base64url wallet signature for the issued challenge."
          }
        },
        "required": [
          "walletAddress",
          "signedChallenge"
        ],
        "example": {
          "walletAddress": "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF",
          "signedChallenge": "vj3mDq2F8wYl9d3r2rYzQhC7Z7Cexample"
        }
      },
      "AuthVerifyResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT bearer token for subsequent API calls."
          }
        },
        "required": [
          "token"
        ],
        "example": {
          "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example"
        }
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "example": {
          "message": "Logged out successfully"
        }
      },
...
      "ManifestRequest": {
        "type": "object",
        "properties": {
          "driverName": {
            "type": "string"
          },
          "driverIdNumber": {
            "type": "string"
          },
          "vehicleRegistration": {
            "type": "string"
          },
          "routeDescription": {
            "type": "string"
          },
          "expectedDeliveryAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "driverName",
          "driverIdNumber",
          "vehicleRegistration",
          "routeDescription",
          "expectedDeliveryAt"
        ]
      },
      "ManifestSubmissionResponse": {
        "type": "object",
        "properties": {
          "manifestId": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ]
          },
          "unsignedXdr": {
            "type": "string"
          }
        },
        "required": [
          "manifestId",
          "unsignedXdr"
        ]
      },
      "ManifestView": {
        "type": "object",
        "additionalProperties": true,
        "description": "Returned shape depends on caller role.\n- Seller views include full manifest fields.\n- Buyer views mask driver identity fields.\n- Mediator/admin views expose hashed identity values.\n",
        "properties": {
          "tradeId": {
            "type": "string"
          },
          "roleView": {
            "type": "string"
          }
        },
        "required": [
          "tradeId"
        ]
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Missing, invalid, revoked, or semantically invalid bearer token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "unauthorized": {
                "value": {
                  "error": "Unauthorized"
                }
              },
              "revoked": {
                "value": {
                  "error": "Unauthorized: token has been revoked"
                }
              }
            }
          }
        }
      },
...
      "NotFoundError": {
        "description": "Requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": "Trade not found"
            }
          }
        }
      },
      "LegacyValidationError": {
        "description": "Request payload failed route-level validation.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              ]
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected backend failure.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                {
                  "$ref": "#/components/schemas/AppErrorResponse"
                }
              ]
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/auth/challenge": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Issue a wallet-signing challenge",
        "description": "Rate limited to 10 attempts per 15-minute window per IP.\nError code references: `VALIDATION_ERROR`, `AUTH_ERROR`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthChallengeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Challenge generated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthChallengeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/LegacyValidationError"
          },
          "429": {
            "description": "Rate limit exceeded for challenge creation.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ]
                },
                "example": "Too many challenges/verify attempts, try again later."
              }
            }
          }
        }
      }
    },
    "/auth/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify the signed challenge and issue a JWT",
        "description": "Rate limited to 10 attempts per 15-minute window per IP.\nError code references: `VALIDATION_ERROR`, `AUTH_ERROR`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signature verified and JWT issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthVerifyResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/LegacyValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "429": {
            "description": "Rate limit exceeded for verification.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ]
                },
                "example": "Too many challenges/verify attempts, try again later."
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Revoke the caller's JWT",
        "description": "Adds the current token `jti` to the revocation denylist until expiry.\nError code references: `AUTH_ERROR`, `INTERNAL_ERROR`.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/trades/{id}/manifest": {
      "get": {
        "tags": [
          "Manifest"
        ],
        "summary": "Fetch the manifest view for a trade",
        "description": "Returned fields depend on caller role:\n- seller receives the full manifest\n- buyer receives masked driver identity\n- mediator/admin receives hashed identity values\nError code references: `AUTH_ERROR`, `DOMAIN_ERROR`, `NOT_FOUND`, `INTERNAL_ERROR`.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TradeIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Manifest returned for the caller's role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManifestView"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": [
          "Manifest"
        ],
        "summary": "Submit a manifest and build the submit-manifest XDR",
        "description": "Error code references: `AUTH_ERROR`, `VALIDATION_ERROR`, `DOMAIN_ERROR`, `NOT_FOUND`, `INTERNAL_ERROR`.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TradeIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManifestRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Manifest accepted and unsigned contract XDR returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManifestSubmissionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/LegacyValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "409": {
            "description": "A manifest already exists for the trade.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  }
}